Refactor: extract tag generation
